Added authentication.
All of the HTML based code seems to work pretty well, and the code seems pretty good, so time to start working on the stuff I started this project for: The REST services.

It's still unclear exactly how well the association is being handled (that is, "Employee.employee_roles") for user accounts that don't have authorization. It's not being shown in the views for display nor update, but I haven't gone through the motions of figuring out if the data can be hacked manually.